requirements.txt
Flask>=2.2,<3.0
line-bot-sdk>=2.4.2
gunicorn>=21.2
python-dotenv>=1.0.1
